REFACTOR: moved api & core files to python pkg
