+++ core/__init__.py
@@ -1,0 +1,2 @@
+from core.core import EDHDeckList, MagicDeckList, MoxFieldUser
+from core.api import MoxFieldAgent
